Update SPI device for T4.x ports are 32 bits
This at least gets the io pins in the ICM2048 test case to output data.

Before the clock pin did not output anything as it is not within the first 8 bits of the port.

This is a more invasive change in that I removed much of the #if code in the main
cpp file and instead created inline helper methods.

Note: the CS versions of these helpers is not used. as the released code did not acutally use the port settings, but did digitalWrite()